Change indexing type from Int64 to Int
Test on x86 failed with complaint about something related to Int32 and
Int64, this commit should probably solve that.

+++ src/read_snapshot/read_format_2.jl
@@ -415,7 +415,7 @@ function read_block_with_offset!(data, n_read::Integer, filename::String, pos0::
     # store position in file
     p = position(f)
 
-    n_read = Int64(n_read) # needed for reading into a view of an array
+    n_read = Int(n_read) # needed for reading into a view of an array
     n_this_key = n_read 
 
     for i = 1:size(offset_key,1)
